refactor: split auth/controller.py + api/controllers/{departments,workflows,providers}.py
